Change Strings/ft_strccmp behaviour
In comparison, the two strings should be treated equally.

ft_strccmp have now the following behaviour : Return the difference
between the two first different characters from s1 and s2, until s1 or
s2 end, c1 is found in s1 or c2 is found in s2.
In the difference, c1 and c2 are considered to be 0.