Remove Types
The Libft must be typesafe. Types shouldn't be dynamics in public
scope. Dynamic types should only be used in private and controlled
scope, for a specific and controlled use.

"sdata" (data with size) does not cause problems semantically, but the
feature isn't used and is too small to exist by itself. it could be
re-implemented (in this form or another) if needed.